sh_status now is a charr unsigned

+++ srcs/builtins/builtin_exit.c
@@ -6,7 +6,7 @@
 /*   By: jmarquet <jmarquet@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/04/07 22:38:07 by jmarquet     #+#   ##    ##    #+#       */
-/*   Updated: 2019/04/07 22:56:16 by jmarquet    ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/04/07 23:09:46 by jmarquet    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -24,5 +24,5 @@ int		builtin_exit(t_sh_state *sh_state, int ac, const char **av, int fd_out)
 	if (ac > 1)
 		return (ft_atoi(av[1]));
 	else
-		return (sh_state->status);
+		return ((int)sh_state->status);
 }
